Consolidate apply_schema_adapter_tests
